feat(chart): Provide ability to use 'full' tag
The [documentation](https://docs.renovatebot.com/getting-started/running/#the-full-image) points users to the 'full' image.
This image is useful when users are behind restrictive firewalls.
However, the chart currently requires users to override the tag to use
the full image. This prevents users from upgrading the image when the
chart upgrades.

This PR provides the ability for users to upgrade the image as and when
chart is upgraded while also using the full image